Bug fixes and additions to Snapdragon2D libraries -Fixed bugs in RenderControl and World2D -Moved KeyEventClient and MouseEventClient to separate compilation units. -Removed Transform from animation library -Added a few more utilitiy methods -Added UI pacakge (work in progress) -Added LICENSE and COPYRIGHT to build JARs
commit build-id=b0000-077

+++ 2DX-GL/snapdragon/com/snap2d/world/World2D.java
@@ -113,8 +113,8 @@ public class World2D {
 		if(ppu < 1)
 			throw(new IllegalArgumentException("illegal pixel-per-unit value: " + ppu));
 		this.ppu = ppu;
-		maxX = minX + ((double)viewWidth / ppu);
-		maxY = minY - ((double)viewHeight / ppu);
+		maxX = minX + (viewWidth / ppu);
+		maxY = minY - (viewHeight / ppu);
 		wt = Math.abs(maxX - minX);
 		ht = Math.abs(maxY - minY);
 	}
