+++ 2DX-GL/x2d/bg/x2d/math/FloatMath.java
@@ -43,7 +43,7 @@ public class FloatMath {
 			return 1;
 		
 		float x = arg0;
-		long lim = (long) Math.abs(power);
+		long lim = Math.abs(power);
 		for(long i = 1; i < lim; i++) {
 			x *= (power > 0) ? arg0:1/arg0;
 		}
